requirements.txt
numpy
pytorch=1.0.1
